clear promotions at begin of testing.

+++ TestPromotion/TestPromotion/Model/Discount.cs
@@ -3,6 +3,6 @@
     public class Discount
     {
         public string Tag;
-        public int Percentage;
+        public decimal Percentage;
     }
 }
